Big simplification!
I'm getting rid of the substitutions; the ability to explicitly fill in
a hole works fine in certain logics, but in general, it cannot be done
soundly, since in (for instance) CTT it would incur further
wellformedness subgoals.

Instead, the setup proposed here is a middle ground between LCF and Coq:
subgoals can express dependencies (in one direction), and as the
synthesis of one subgoal is filled in, it is substituted into the
subgoals to the right of it.