Change the semantics of SEQ a little
This change is kind of nice; whereas previously, binding hyps was like
pushing them onto a stack for the rest of the script, now it only pushes
them onto the stack *for the duration* of the first argument to SEQ.

So, now the behavior of the following scripts is different:

    a, b, c <- (tac1; tac2);
    tac3

vs

    a, b, c <- tac1;
    tac2;
    tac3

Previously, [a,b,c] would in either case be consumed by tac1, tac2, tac3.

Now, in the former script, they are consumed by both (tac1, tac2), and
not by tac3; in the latter script, they are consumed only by tac1,
whereas tac2, tac3 do not consume those names.